pyproject.toml
[project]
name = "flow-metric-service"
version = "0.1.0"
description = "Microservice for computing flow metric anomaly scores for IFCB bins."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "amplify-stateless @ git+https://github.com/WHOIGit/amplify-stateless-microservice.git@v1.3.1",
    "amplify-storage-utils @ git+https://github.com/WHOIGit/amplify-storage-utils.git@v1.3.2",
    "pyifcb @ git+https://github.com/joefutrelle/pyifcb.git@v1.2.1",
    "ifcb-flow-metric @ git+https://github.com/WHOIGit/ifcb-flow-metric.git@v1.0.2",
    "scikit-learn",
    "pandas",
    "numpy",
    "PyYAML",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["flow_metric_service"]

[tool.hatch.metadata]
allow-direct-references = true
